+++ api/composer.json
@@ -1,0 +1,5 @@
+{
+    "require": {
+        "stripe/stripe-php": "^14.6"
+    }
+}
