+++ package.json
@@ -29,6 +29,7 @@
   },
   "scripts": {
     "start": "react-scripts start",
+    "start-backend": "php -S localhost:8000 -t ../api",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
